ISSUE #4321 - FIxed a couple of circular dependencies with stores and groups helpers
